refactor: relocate files into appropriate packages
